Moved integration tests to it folder
