+++ .travis.yml
@@ -1,25 +1,26 @@
 branches:
   only:
-  - master
-language: scala
-scala:
-- 2.10.4
-- 2.11.7
-jdk:
-- openjdk7
-- oraclejdk8
-services:
-- cassandra
-cache:
-  directories:
-  - "~/.ivy2"
-  apt: true
-sbt_args: "-Dsbt.log.noformat=true -sbt-dir ~/.sbt"
-after_success:
-- sh "$TRAVIS_BUILD_DIR/.travis.release.sh"
+    - master
 env:
   global:
-  - secure: VoNWw9dVz5fugaJzyZ5LgZY0B7OOwehwZZ/iuZni0b/FISZJIY/CXdpSdflB/+EsX2zUPQseGNvjdfPHyoc7vlDVv9bNTSSIncgiElLXeSV1cgOcnelXd/aPJiGB1ITlg0x1Iej2SB5E7AE9/+DoTBXUtXdjnqAdfXMG6ePb/Z6mC1QAFu5nC8CGQy2fIxG0dF83fq+HP6QAYKsgpqnl4XJ5XO7bMxbjhUFBgWLd4Y/gO902D+D82Yjems/wgtfztSzg9AxRLMCiTEzxs8Q87ZduIKd11MIkkGUGf9o6T1Hty8mm9qcgLxu8HkiCia6LzHv9Bpyx1A6wDXchyfy9iR+7V+07JlLILgEKKh1KuOgmlCCC+UiLHeXjKAgXeiW3WpC7gzTADQtL0Y3+9qA9WZ67xpkaPVBQPJF3vE59ivmPwxYw76w+UIRx16lSEhkg7A80lz5At9nOpTXzXA6L1fuIYlIWb4W9AYPlZSatP+IOJ4QcyH0aupXmWTMMY1U54W4WqlVBqyEeb1IfJfggyWjdWZ2pk4woHxgntvgshIruu3wRp2iLqfYnH3thLiWxvcu4BAeO8GaXikfefvKzvgLDotxekJwbLTPpP7GycYAAlTs3cdvP+mDIBCfRsMvP4fEJoBRoCYE6v1gweU9uKT3bPNvGok6a+x5lovnyFYE=
-  - secure: d9SKG2KrX8EGilXRttUHQ7Lu1wmBflXSHmnDgP2DO4RZSiUkSzNK793JT85m/hzUKIPs+l3uLsKWYI4ki1H/vHnM4i7ofk4mc6Xb+xpubVFCSdX3pBhF66HlUxhBcS9zGNrKZHkypqeGLrxyx4GZfR+Ghd4vaJIZQNq/UuS8WjnUvIfkDffxm3rdm5YcR9JHk5GGp4uuqLzhYXhUsHTNk/xQWmC2/ryv0l4jV6rPVkxePRC5bGmbJJNetWtsVHj0+aeJ/q0oKjHFCvUcBCBfe2eulwIbLqgE7SEOsujaXJcm0PVcMRJY3dRHGOQGbVyMxzRxgnJ0RnGB23abKJDvt3Y5y1Lc9oJL9MLQtlXdjAiFNb0A1jxUwkHrHYkBoHVlKg30i5wCbocqSahahN/bY4ryWG0iG+e0IEcS4ZbAA7+CpGxj+56XrcbSvG4G2rfiZ4DZlLt6zJnV+pGLowTmbxjkf3PqAHQUpolE0y5GQJw3Y3h/1OJZvTu+eZzuolJLgAbPriB5ZuCkJg78XCeVBXwHdeFUxcSPtaKt4MvEgRNEUzBKvQJP2Ivh0QVh6W4UOi6DmU97nBFdZi5O/WBPVokXzKdCkooi8RE238a8gbnWGhYkV8oZOV/ORldnJgZBLEEDSqIcDxkLMuhVEwbYF0NTLo3oSksycL7NoBplLqc=
-  - secure: WQjvevUqSHRpb3T5pgFD5Ci1U0tM30xyAQcA1x5A9W2Fmkmc+g3IkcJfEXePnk5+Z93LVikuiq6vlHXryPN4J9gS55VCiv5ciqHXN3Yk4XCe+/mtWAb+HY0EVou9abRdiicLHqOvXgT9Ijd9oUpdOsF1rjR0NQq0AQgX59uTi2Zi8iTEu5QQSCIOllmYtivG7h9gkR3fKqjP/GwsPBGmgLdG/DDUCe/pkVRZf/78pXroaohDZyuyhlnD9+asXHaxQ7U0u918dq5mzDccNBoG3tdaR22mppPWZbBiamRaUejm+D8yxzCe4Q2lvwczGIEQxyNtTsHZAsQSZJB96+tm81t5ncmJhwBVXclf85g7znnaVnrN0hiR2f4e6B4aCnpBpEm2/y+IuLH6cDe+0ikCapcOjFKUPFIgQcwPmXv7baGFfgvB6ZJrCnfR5MIGRUn0PguYFH1BZcGABcPkW+PexmMBU5kD4MfINd5cal0z6QANaD/teG1VuEwAM8+Td/iLrhrqyLl9YRJXXEdAiib5+iCxeiTA8lf4H9RER40XmVUwFpm3CKBkpoJWbdw5bi89E4+8wXsQQ24IiRv27Baa0ZED1X4kd2XwlbBp6en0ADp2QKyx0TySTwB9YjJyy84e27+I12uaSKjvdv8OVVfnxqcxmEzrO+TsMO8jp82sO4k=
-  - secure: PK/YBp8YTbtL+9XcSGaCejjIVkaMwlIe0CGOTFfP1DoHr3W+ipLD/SBh51QeqJwIuSlAUzL8GaHECwYDzJR5YK1P922L9+GzNTTFBLqyFa1gJPibp4GwPcCLYzw9n5/Raj9sMR2LJ2VtvPbhZHZlNPJvVIrYppjDMdDmM6Fsc+M3eJRjnpI0nNRUD26Gx7zD7yuWffjSemAJGWFvaEhH6iQYun9cHzmRH1rHDUpM7oqaGLFkNvOTHZRZGQlkBNCmc+IjtdlC4cgsxYq12y2y4PCi7ec7TQ6VwDGCOQPdGSay8jMrSTBNVVoEI9RjkaEV0ITJt2mDmUTvC3jzKtKCEoZJGxKtkpF65ltMMp1a7pwo26v1sJ4m7vuKA+8+UVSndEFkRRhFB5Usfkvd7fFMG2ezlVTdN99rH/2y3NHdGCRWuEqeZzho25H0V6h1psS+7ZDCn5nahU5oPoS/ChzvwW1Oc2Z5NmwP57sGBz+QKmQ6qIlLLfh7b32TAzBA5a82YIVtA1gpin+g0Dyr9/cn7A3lrK09l4OABQ1edI4eoXaiHsNBGrjNtpjbi8vNyaTfY0vm0SjgKuqoJf6UIJssb9zWZ0Gf9J6GSIuYTUHZgTaLTTtL5gK61MTIq4L9k7dQmnoNGdC6Lxh1X3MJBMP7yAYrnaDd4iAZZEMyCgequ38=
+    - secure: VoNWw9dVz5fugaJzyZ5LgZY0B7OOwehwZZ/iuZni0b/FISZJIY/CXdpSdflB/+EsX2zUPQseGNvjdfPHyoc7vlDVv9bNTSSIncgiElLXeSV1cgOcnelXd/aPJiGB1ITlg0x1Iej2SB5E7AE9/+DoTBXUtXdjnqAdfXMG6ePb/Z6mC1QAFu5nC8CGQy2fIxG0dF83fq+HP6QAYKsgpqnl4XJ5XO7bMxbjhUFBgWLd4Y/gO902D+D82Yjems/wgtfztSzg9AxRLMCiTEzxs8Q87ZduIKd11MIkkGUGf9o6T1Hty8mm9qcgLxu8HkiCia6LzHv9Bpyx1A6wDXchyfy9iR+7V+07JlLILgEKKh1KuOgmlCCC+UiLHeXjKAgXeiW3WpC7gzTADQtL0Y3+9qA9WZ67xpkaPVBQPJF3vE59ivmPwxYw76w+UIRx16lSEhkg7A80lz5At9nOpTXzXA6L1fuIYlIWb4W9AYPlZSatP+IOJ4QcyH0aupXmWTMMY1U54W4WqlVBqyEeb1IfJfggyWjdWZ2pk4woHxgntvgshIruu3wRp2iLqfYnH3thLiWxvcu4BAeO8GaXikfefvKzvgLDotxekJwbLTPpP7GycYAAlTs3cdvP+mDIBCfRsMvP4fEJoBRoCYE6v1gweU9uKT3bPNvGok6a+x5lovnyFYE=
+    - secure: d9SKG2KrX8EGilXRttUHQ7Lu1wmBflXSHmnDgP2DO4RZSiUkSzNK793JT85m/hzUKIPs+l3uLsKWYI4ki1H/vHnM4i7ofk4mc6Xb+xpubVFCSdX3pBhF66HlUxhBcS9zGNrKZHkypqeGLrxyx4GZfR+Ghd4vaJIZQNq/UuS8WjnUvIfkDffxm3rdm5YcR9JHk5GGp4uuqLzhYXhUsHTNk/xQWmC2/ryv0l4jV6rPVkxePRC5bGmbJJNetWtsVHj0+aeJ/q0oKjHFCvUcBCBfe2eulwIbLqgE7SEOsujaXJcm0PVcMRJY3dRHGOQGbVyMxzRxgnJ0RnGB23abKJDvt3Y5y1Lc9oJL9MLQtlXdjAiFNb0A1jxUwkHrHYkBoHVlKg30i5wCbocqSahahN/bY4ryWG0iG+e0IEcS4ZbAA7+CpGxj+56XrcbSvG4G2rfiZ4DZlLt6zJnV+pGLowTmbxjkf3PqAHQUpolE0y5GQJw3Y3h/1OJZvTu+eZzuolJLgAbPriB5ZuCkJg78XCeVBXwHdeFUxcSPtaKt4MvEgRNEUzBKvQJP2Ivh0QVh6W4UOi6DmU97nBFdZi5O/WBPVokXzKdCkooi8RE238a8gbnWGhYkV8oZOV/ORldnJgZBLEEDSqIcDxkLMuhVEwbYF0NTLo3oSksycL7NoBplLqc=
+    - secure: WQjvevUqSHRpb3T5pgFD5Ci1U0tM30xyAQcA1x5A9W2Fmkmc+g3IkcJfEXePnk5+Z93LVikuiq6vlHXryPN4J9gS55VCiv5ciqHXN3Yk4XCe+/mtWAb+HY0EVou9abRdiicLHqOvXgT9Ijd9oUpdOsF1rjR0NQq0AQgX59uTi2Zi8iTEu5QQSCIOllmYtivG7h9gkR3fKqjP/GwsPBGmgLdG/DDUCe/pkVRZf/78pXroaohDZyuyhlnD9+asXHaxQ7U0u918dq5mzDccNBoG3tdaR22mppPWZbBiamRaUejm+D8yxzCe4Q2lvwczGIEQxyNtTsHZAsQSZJB96+tm81t5ncmJhwBVXclf85g7znnaVnrN0hiR2f4e6B4aCnpBpEm2/y+IuLH6cDe+0ikCapcOjFKUPFIgQcwPmXv7baGFfgvB6ZJrCnfR5MIGRUn0PguYFH1BZcGABcPkW+PexmMBU5kD4MfINd5cal0z6QANaD/teG1VuEwAM8+Td/iLrhrqyLl9YRJXXEdAiib5+iCxeiTA8lf4H9RER40XmVUwFpm3CKBkpoJWbdw5bi89E4+8wXsQQ24IiRv27Baa0ZED1X4kd2XwlbBp6en0ADp2QKyx0TySTwB9YjJyy84e27+I12uaSKjvdv8OVVfnxqcxmEzrO+TsMO8jp82sO4k=
+    - secure: PK/YBp8YTbtL+9XcSGaCejjIVkaMwlIe0CGOTFfP1DoHr3W+ipLD/SBh51QeqJwIuSlAUzL8GaHECwYDzJR5YK1P922L9+GzNTTFBLqyFa1gJPibp4GwPcCLYzw9n5/Raj9sMR2LJ2VtvPbhZHZlNPJvVIrYppjDMdDmM6Fsc+M3eJRjnpI0nNRUD26Gx7zD7yuWffjSemAJGWFvaEhH6iQYun9cHzmRH1rHDUpM7oqaGLFkNvOTHZRZGQlkBNCmc+IjtdlC4cgsxYq12y2y4PCi7ec7TQ6VwDGCOQPdGSay8jMrSTBNVVoEI9RjkaEV0ITJt2mDmUTvC3jzKtKCEoZJGxKtkpF65ltMMp1a7pwo26v1sJ4m7vuKA+8+UVSndEFkRRhFB5Usfkvd7fFMG2ezlVTdN99rH/2y3NHdGCRWuEqeZzho25H0V6h1psS+7ZDCn5nahU5oPoS/ChzvwW1Oc2Z5NmwP57sGBz+QKmQ6qIlLLfh7b32TAzBA5a82YIVtA1gpin+g0Dyr9/cn7A3lrK09l4OABQ1edI4eoXaiHsNBGrjNtpjbi8vNyaTfY0vm0SjgKuqoJf6UIJssb9zWZ0Gf9J6GSIuYTUHZgTaLTTtL5gK61MTIq4L9k7dQmnoNGdC6Lxh1X3MJBMP7yAYrnaDd4iAZZEMyCgequ38=
+language: scala
+scala:
+  - 2.10.4
+  - 2.11.7
+jdk:
+  - openjdk7
+  - oraclejdk8
+services:
+  - cassandra
+cache:
+  directories:
+    - "~/.ivy2"
+  apt: true
+script:
+  - sbt -Dsbt.log.noformat=true -sbt-dir ~/.sbt ++$TRAVIS_SCALA_VERSION test it:test
+after_success:
+  - sh "$TRAVIS_BUILD_DIR/.travis.release.sh"
